requirements.txt
streamlit>=1.22.0
python-pptx>=0.6.21
requests>=2.28.1
python-dotenv>=0.21.0